feat(NoNoticeScreenContrib): get value
