+++ src/app/content/App/Contribute/ContributeScreen/SubmitContributionForm/withReduxForm.ts
@@ -1,9 +1,17 @@
-import { reduxForm } from 'redux-form';
+import { formValueSelector, reduxForm } from 'redux-form';
+import { connect } from 'react-redux';
+import { ContentState } from 'app/content/store/reducers';
 import validate from 'libs/domain/contribution/validateForm';
 import { Contribution } from 'libs/domain/notice';
 import { SubmitContributionFormOwnProps } from './SubmitContributionForm';
 
 export const form = 'contribution';
+
+const getMessageValue = (state: ContentState) => ({
+  message: formValueSelector(form)(state, 'message')
+});
+
+export const withMessageValue = connect(getMessageValue);
 
 export default reduxForm<Contribution, SubmitContributionFormOwnProps>({
   form,
