+++ src/app/profiles/store/sagas/index.ts
@@ -7,7 +7,8 @@ import subscriptionsSaga from 'app/profiles/store/sagas/subscriptions.saga';
 import categoriesSaga from 'app/profiles/store/sagas/categories.saga';
 import locationChangeSaga from './locationChange.saga';
 import contributorSaga from './contributor.saga';
-
+// eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+// @ts-ignore
 const connectionSaga = browser?.runtime?.connect
   ? portConnectionSaga
   : windowConnectionSaga.bind(null, process.env.PROFILES_ORIGIN);
